Development of project module at the expense of acquisition. Layout editor development. Work-around for Qt "Bug" finding a QLabel buddy by name loading ui-file twice using FormBuilder. Added layout editor property page for changing the Widget's assigned layout. PropertyPage canApply() added. Development AscanGraph and property page.

+++ com/rsa/iface/RsaTypes.h
@@ -25,7 +25,7 @@ struct _RSA_CLASS RsaTypes
 	/**
 	 * @brief Type for gathering parameter ID's.
 	 */
-	typedef uint64_t IdType;
+	typedef unsigned long long IdType;
 
 	/**
 	 * @brief Type for gathering parameter ID's.
